fix: don't append zero time as NULL without nullzero tag
