+++ dialect/append.go
@@ -95,9 +95,6 @@ func AppendBytes(b []byte, bytes []byte) []byte {
 }
 
 func AppendTime(b []byte, tm time.Time) []byte {
-	if tm.IsZero() {
-		return AppendNull(b)
-	}
 	b = append(b, '\'')
 	b = tm.UTC().AppendFormat(b, "2006-01-02 15:04:05.999999-07:00")
 	b = append(b, '\'')
