+++ dialect/mysqldialect/append.go
@@ -31,9 +31,6 @@ func appendTimeValue(fmter schema.Formatter, b []byte, v reflect.Value) []byte {
 }
 
 func appendTime(b []byte, tm time.Time) []byte {
-	if tm.IsZero() {
-		return dialect.AppendNull(b)
-	}
 	b = append(b, '\'')
 	b = tm.UTC().AppendFormat(b, "2006-01-02 15:04:05.999999")
 	b = append(b, '\'')
