+++ internal/dbtest/testdata/snapshots/TestQuery-mysql8-84
@@ -1,0 +1,1 @@
+INSERT INTO `models` (`time`) VALUES ('0001-01-01 00:00:00')
